pyproject.toml
[project]
name = "nebula-crm"
version = "0.0.0"
description = "Nebula CRM tooling dependencies"
requires-python = ">=3.12"
dependencies = [
  "pyyaml==6.0.2",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
